requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
python-multipart
python-dotenv
httpx
redis
rq
jwt